737/745

+++ libft/ft_putnbr_fd.c
@@ -6,15 +6,15 @@
 /*   By: atweek <atweek@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/13 15:05:51 by atweek            #+#    #+#             */
-/*   Updated: 2021/01/18 11:18:29 by atweek           ###   ########.fr       */
+/*   Updated: 2021/01/23 04:54:20 by atweek           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putnbr_fd(int nb, int fd)
+void	ft_putnbr_fd(long long nb, int fd)
 {
-	unsigned int	nbr;
+	unsigned long long	nbr;
 
 	if (nb < 0)
 	{
